access denied

+++ src/main/java/com/mugi/logicea/cards/service/CardService.java
@@ -1,13 +1,16 @@
 package com.mugi.logicea.cards.service;
 
-import com.mugi.logicea.cards.dtos.CreateCardRequest;
 import com.mugi.logicea.cards.dtos.CardStatus;
+import com.mugi.logicea.cards.dtos.CreateCardRequest;
 import com.mugi.logicea.cards.dtos.UpdateCardRequest;
 import com.mugi.logicea.cards.mapper.CardMapper;
 import com.mugi.logicea.entiy.Card;
 import com.mugi.logicea.repository.CardRepository;
-import com.mugi.logicea.utils.*;
-import jakarta.validation.ConstraintViolation;
+import com.mugi.logicea.uac.dtos.Role;
+import com.mugi.logicea.utils.ErrorMessage;
+import com.mugi.logicea.utils.GenericSpecifications;
+import com.mugi.logicea.utils.RestResponseObject;
+import com.mugi.logicea.utils.SearchDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ObjectUtils;
@@ -19,7 +22,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
-import java.util.*;
+import java.util.Date;
+import java.util.List;
+import java.util.Optional;
 
 import static com.mugi.logicea.utils.UtilFunctions.*;
 
@@ -30,7 +35,7 @@ public class CardService {
     private final GenericSpecifications<Card> cardSpecification = new GenericSpecifications<>();
     private final CardRepository cardRepository;
 
-    public ResponseEntity<RestResponseObject<Card>>  createCard(CreateCardRequest request) {
+    public ResponseEntity<RestResponseObject<Card>> createCard(CreateCardRequest request) {
         Card card = CardMapper.INSTANCE.createRequestToCardEntity(request);
         List<ErrorMessage> validateObj = validateNotNull(request);
         if (ObjectUtils.isNotEmpty(validateObj)) {
@@ -45,15 +50,14 @@ public class CardService {
         card.setModifiedDate(Date.from(Instant.now()));
         card.setCreateBy(currentUser);
         card.setModifiedBy(currentUser);
-        cardRepository.save(card);
+        Card newCard = cardRepository.save(card);
         return new ResponseEntity(
-                RestResponseObject.builder().message("Card Created Successfully").build(),
+                RestResponseObject.builder().message("Card Created Successfully").payload(newCard).build(),
                 HttpStatus.OK);
     }
 
 
-
-    public ResponseEntity<RestResponseObject<Card>>  updateCard(Long id, UpdateCardRequest request) {
+    public ResponseEntity<RestResponseObject<Card>> updateCard(Long id, UpdateCardRequest request) {
         List<ErrorMessage> validateObj = validateNotNull(request);
         if (ObjectUtils.isNotEmpty(validateObj)) {
 
@@ -68,11 +72,16 @@ public class CardService {
             card.setCreateDate(existingCard.get().getCreateDate());
             card.setCreateBy(existingCard.get().getCreateBy());
             var currentUser = getCurrentUser();
+            if (ObjectUtils.notEqual(currentUser, existingCard.get().getCreateBy())) {
+                return new ResponseEntity(
+                        RestResponseObject.builder().message("Access denied!!").build(),
+                        HttpStatus.BAD_REQUEST);
+            }
             card.setModifiedDate(Date.from(Instant.now()));
             card.setModifiedBy(currentUser);
-            cardRepository.save(card);
+            Card updatedCard = cardRepository.save(card);
             return new ResponseEntity(
-                    RestResponseObject.builder().message("Card Updated Successfully").build(),
+                    RestResponseObject.builder().message("Card Updated Successfully").payload(updatedCard).build(),
                     HttpStatus.OK);
         } else {
             return new ResponseEntity(
@@ -82,7 +91,7 @@ public class CardService {
         }
     }
 
-    public ResponseEntity<RestResponseObject<Card>>  fetchCard(SearchDto searchDto) {
+    public ResponseEntity<RestResponseObject<Card>> fetchCard(SearchDto searchDto) {
         List<ErrorMessage> validateObj = validateNotNull(searchDto);
         if (ObjectUtils.isNotEmpty(validateObj)) {
             return new ResponseEntity(
@@ -92,12 +101,17 @@ public class CardService {
         try {
             Specification<Card> cardSpec = cardSpecification.findByFieldAndValue(searchDto);
             Optional<Card> card = cardRepository.findOne(cardSpec);
-            if (card.isEmpty()) new ResponseEntity(
+            if (card.isEmpty()) return new ResponseEntity(
                     RestResponseObject.builder().message("No card found!!").build(),
                     HttpStatus.NOT_FOUND);
+            if (validateUserHasNoAccess(card.get().getCreateBy())) {
+                return new ResponseEntity(
+                        RestResponseObject.builder().message("Access denied!!").build(),
+                        HttpStatus.BAD_REQUEST);
+            }
             return new ResponseEntity(
                     RestResponseObject.builder().message("Fetched Successfully")
-                            .payload(card)
+                            .payload(card.get())
                             .build(),
                     HttpStatus.OK);
         } catch (Exception e) {
@@ -110,9 +124,17 @@ public class CardService {
     public ResponseEntity<Page<Card>> fetchPaginatedCardList(SearchDto searchDto, Pageable pageable) {
         try {
             Specification<Card> cardSpec = Specification.where(null);
+
             if (ObjectUtils.isNotEmpty(searchDto.getFieldName())
                     && ObjectUtils.isNotEmpty(searchDto.getValue())) {
                 cardSpec = cardSpecification.findByFieldAndValue(searchDto);
+            }
+            String userRole = getCurrentUserRole();
+            if (userRole.equals(Role.MEMBER.toString())) {
+                cardSpec = cardSpec.and(cardSpecification.findByFieldAndValue(SearchDto.builder()
+                        .fieldName("createBy")
+                        .value(getCurrentUser())
+                        .build()));
             }
             Page<Card> page = cardRepository.findAll(cardSpec, pageable);
 
@@ -130,6 +152,12 @@ public class CardService {
 
     public ResponseEntity<RestResponseObject> deleteCard(Long id) {
         Optional<Card> existingCard = cardRepository.findById(id);
+        var currentUser = getCurrentUser();
+        if (ObjectUtils.notEqual(currentUser, existingCard.get().getCreateBy())) {
+            return new ResponseEntity(
+                    RestResponseObject.builder().message("Access denied!!").build(),
+                    HttpStatus.BAD_REQUEST);
+        }
         if (existingCard.isPresent()) {
 
             cardRepository.delete(existingCard.get());
@@ -142,5 +170,13 @@ public class CardService {
                     HttpStatus.NOT_FOUND);
 
         }
+    }
+
+    boolean validateUserHasNoAccess(String createdBy) {
+        String userRole = getCurrentUserRole();
+        if (ObjectUtils.equals(userRole, Role.ADMIN.toString())) return false;
+        var currentUser = getCurrentUser();
+        return ObjectUtils.notEqual(currentUser, createdBy);
+
     }
 }
